src/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
astroid==3.2.4
attrs==23.2.0
colorama==0.4.6
dill==0.3.8
docopt==0.6.2
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
iso8601==2.1.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
Pygments==2.18.0
pylint==3.2.6
pytest==8.3.2
pytest-watch==4.2.0
python-dotenv==1.0.1
rich==13.7.1
tomlkit==0.13.0
twitchio==2.10.0
typing_extensions==4.12.2
watchdog==4.0.1
yarl==1.9.4
